feat: 요청유저와 pathVariable 유저 같은지 검증

+++ dg/src/main/java/com/dajungdagam/dg/service/UserService.java
@@ -90,12 +90,15 @@ public class UserService {
 
     // 유저 별명 업데이트
     @Transactional
-    public int updateUserNickName(int userId, String nickName) {
+    public int updateUserNickName(UserResponseDto userResponseDto, int userId, String nickName) {
         int id = -1;
         try{
 
             Optional<User> userObj = repository.findById(userId);
             User user = userObj.get();
+
+            // 요청 보낸 유저와 같은지 체크
+            if(!isSameUser(userId, userResponseDto))   throw new Exception("권한이 없습니다.");
 
             user.setNickName(nickName);
 
@@ -113,12 +116,15 @@ public class UserService {
     // 유저 사는곳 업데이트
 
     @Transactional
-    public int updateUserArea(int userId, String gu, String dong){
+    public int updateUserArea(UserResponseDto userResponseDto, int userId, String gu, String dong){
         int id = -1;
         try{
 
             Optional<User> userObj = repository.findById(userId);
             User user = userObj.get();
+
+            // 요청 보낸 유저와 같은지 체크
+            if(!isSameUser(userId, userResponseDto))   throw new Exception("권한이 없습니다.");
 
             Area area = areaRepository.findByGuNameAndDongName(gu, dong);
             user.setArea(area);
@@ -133,12 +139,15 @@ public class UserService {
     }
 
     @Transactional
-    public int updateUserInfo(int userId, String info) {
+    public int updateUserInfo(UserResponseDto userResponseDto, int userId, String info) {
         int id = -1;
         try{
 
             Optional<User> userObj = repository.findById(userId);
             User user = userObj.get();
+
+            // 요청 보낸 유저와 같은지 체크
+            if(!isSameUser(userId, userResponseDto))   throw new Exception("권한이 없습니다.");
 
             user.setInfo(info);
 
@@ -153,10 +162,13 @@ public class UserService {
 
 
     @Transactional
-    public boolean deleteUser(int userId) {
+    public boolean deleteUser(UserResponseDto userResponseDto, int userId) {
         try {
             Optional<User> userObj = repository.findById(userId);
             User user = userObj.get();
+
+            // 요청 보낸 유저와 같은지 체크
+            if(!isSameUser(userId, userResponseDto))   throw new Exception("권한이 없습니다.");
 
             boolean res = postService.deleteAllPost(user);
             if (!res) throw new Exception("User delete failed!");
@@ -184,9 +196,10 @@ public class UserService {
 //        return true;
 //    }
 
-    public boolean isSameUser(int userId, UserResponseDto userResponseDto){
-        User user = userResponseDto.getUser();
 
+
+    public static boolean isSameUser(int userId, UserResponseDto userResponseDto){
+        User user = userResponseDto.getUser();
 
         return user.getId() == userId;
     }
